+++ apps/Diary/serializers/diary_group_member_sz.py
@@ -1,8 +1,8 @@
 from rest_framework.serializers import ModelSerializer
-from ..models import Diary, DiaryGroup,DiaryMember
-from .diary_list_sz import DiaryListSZ
+from ..models import DiaryGroupMember
+
 
 class DiaryGroupMemberListSZ(ModelSerializer):
     class Meta:
-        model = DiaryMember
-        fields = ['id', 'rank', 'diary','created_at', 'updated_at']
+        model = DiaryGroupMember
+        fields = ['id', 'rank', 'group', 'diary', 'created_at', 'updated_at']
